% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_patients}
\alias{add_patients}
\title{Add all included patients to data frame}
\usage{
add_patients(tidy, patients)
}
\arguments{
\item{tidy}{data_frame with the data being returned}

\item{patients}{data_frame with a column \code{pie.id} that contains all
included patients}
}
\value{
data_frame
}
\description{
Makes sure all patients are included in the data frame and the values for
added patients are \code{FALSE}
}
\keyword{internal}

