% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{assign_class}
\alias{assign_class}
\title{Assign edwr class type}
\usage{
assign_class(df, x, new_class)
}
\arguments{
\item{df}{new data_frame}

\item{x}{initial data_frame}

\item{new_class}{string with new class name}
}
\value{
data frame
}
\description{
Assign edwr class type
}
\keyword{internal}
