% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{assign_names}
\alias{assign_names}
\title{Assign desired names to columns}
\usage{
assign_names(x, varnames, extras = NULL)
}
\arguments{
\item{x}{data frame}

\item{varnames}{named list of standard column names for given class}

\item{extras}{optional named list of additional column names}
}
\value{
data frame
}
\description{
Assign desired names to columns
}
\keyword{internal}
