% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_perctime.R
\name{calc_perctime}
\alias{calc_perctime}
\alias{calc_perctime.default}
\alias{calc_perctime.meds_cont}
\alias{calc_perctime.meds_inpt}
\alias{calc_perctime.meds_sched}
\alias{calc_perctime.events}
\alias{calc_perctime.labs}
\alias{calc_perctime.vitals}
\title{Calculate proportion of time above or below a threshold}
\usage{
calc_perctime(x, ...)

\method{calc_perctime}{default}(x, ...)

\method{calc_perctime}{meds_cont}(x, thrshld, ...)

\method{calc_perctime}{meds_inpt}(x, thrshld, cont = TRUE, ...)

\method{calc_perctime}{meds_sched}(x, thrshld, ...)

\method{calc_perctime}{events}(x, thrshld, ...)

\method{calc_perctime}{labs}(x, thrshld, ...)

\method{calc_perctime}{vitals}(x, thrshld, ...)
}
\arguments{
\item{x}{A data frame with serial measurement data}

\item{...}{additional arguments passed on to individual methods}

\item{thrshld}{A charactor vector of the criteria}

\item{cont}{A logical, if TRUE (default), treat the data as continuous
when calculating the percent time}
}
\value{
A data frame
}
\description{
\code{calc_perctime} calculates percent time above / below a threshold
}
\details{
This function takes a data frame with serial measurement data and produces a
data frame with percent time above or below a threshold for each infusion.

The argument \code{thrshld} must be provided as a list with one or more
criteria, which indicate whether to calculate time above, below, or between
the threshold values. The criteria should be written as a formula, of the
form: \code{list(~variable > value)} or \code{list(~variable < value1,
~variable > value2)}
}
\examples{
# make a reference data frame for tidying meds
library(dplyr)
ref <- tibble::tibble(
  name = c("heparin", "warfarin", "antiplatelet agents"),
  type = c("med", "med", "class"),
  group = c("cont", "sched", "sched")
)

# tidy continuous medications; will keep only heparin drips
x <- tidy_data(meds_cont, meds_sched, ref)
y <- calc_runtime(x)

# calculate the proportion of time the infusion rate was > 10 units/kg/hour
print(head(
  calc_perctime(y, "med.rate > 10")
))

}
