% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runtime.R
\name{calc_runtime_fun}
\alias{calc_runtime_fun}
\title{Performs the runtime calculation}
\usage{
calc_runtime_fun(x, ..., val_col, dt_col, units = "hours")
}
\arguments{
\item{x}{tibble}

\item{...}{grouping quosures}

\item{val_col}{event column (lab, med, etc.)}

\item{dt_col}{datetime column}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}
}
\value{
tibble
}
\description{
Performs the runtime calculation
}
\keyword{internal}
