% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crcl.R
\name{calculate_crcl}
\alias{calculate_crcl}
\title{Caculate the creatinine clearance}
\usage{
calculate_crcl(age, sex, scr, weight, height)
}
\arguments{
\item{age}{A numeric with the patient's age}

\item{sex}{A character indicating patient's gender, as "Male" or "Female"}

\item{scr}{A numeric with the serum creatinine, in mg/dL}

\item{weight}{A numeric with the patient's actual body weight, in kg}

\item{height}{A numeric with the patient's height, in cm}
}
\value{
A numeric, creatinine clearance in mL/min
}
\description{
\code{calculate_crcl} calculates the creatinine clearance using the
Cockcroft-Gault equation
}
\details{
This function calculates the creatinine clearance for a patient based on the
Cockcroft-Gault equation: CrCl[mL/min] = Sex * ((140 - Age[yr]) / Serum
Creatinine[mg/dL]) * (Lean Body Weight[kg] / 72); where Sex = 1 for Males and Sex =
0.85 for Females.
}
\references{
Cockcroft DW, Gault MH. Prediction of creatinine clearance from serum creatinine. Nephron. 1976;16(1):31-41
}
