% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_crcl.R
\name{calculate_leanbw}
\alias{calculate_leanbw}
\title{Calculate the Lean Body Weight}
\usage{
calculate_leanbw(sex, height)
}
\arguments{
\item{sex}{A character indicating patient's gender, as "Male" or "Female"}

\item{height}{A numeric with the patient's height, in cm}
}
\value{
A numeric, ideal body weight, in kg
}
\description{
\code{calculate_leanbw} calculates the ideal body weight in kg
}
\details{
This function takes a patient's height and based on their gender returns the
patient's lean body weight in kg. For males, the equation is: LBW[kg] = (0.73
* Height[cm]) - 59.42; for females: LBW[kg] = (0.65 * Height[cm]) - 50.74.
}
\references{
Burton ME, Chow MS, Platt DR, et. al. Accuracy of Bayesian and Sawchuk-Zaske dosing methods for gentamicin. Clin Pharm. 1986 Feb;5(2):143-9
}
