% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{check_rename}
\alias{check_rename}
\title{Check if column names have already been renamed by read_data2}
\usage{
check_rename(x)
}
\arguments{
\item{x}{data frame}
}
\value{
logical
}
\description{
Check if column names have already been renamed by read_data2
}
\keyword{internal}
