% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnosis.R
\docType{data}
\name{diagnosis}
\alias{diagnosis}
\title{Sample diagnosis data}
\format{A data frame with sample diagnosis data for 15 patients. The data has
  been read in via \code{\link{read_data}} but no further tidying or
  transformation has been performed.

  Note that the present on admission field does not seem to be popluated in
  the EDW.
\describe{
  \item{pie.id}{encounter identifier}
  \item{diag.code}{ICD-9/10-CM code}
  \item{code.source}{type of ICD code}
  \item{diag.type}{status of the coding (e.g., Admitting, Final)}
  \item{diag.seq}{primary or secondary diagnosis indicator}
  \item{present.admit}{flag for disease present on admission}
}}
\usage{
diagnosis
}
\description{
A dataset containing sample ICD-9-CM and ICD-10-CM codes
}
\keyword{datasets}

