% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edwr.R
\docType{package}
\name{edwr}
\alias{edwr}
\alias{edwr-package}
\title{Standardized Functions for Working with Data from the EDW}
\description{
\pkg{edwr} provides functions to facilitate working with data extracted from
the EDW
}
\details{
This package provides standardized functions for reading, tidying, and
transforming data which was extracted from the Enterprise Data Warehouse
(EDW) using the standardized EDW queries.

The main functions include:
\enumerate{
  \item \code{\link{read_data}} for reading data from csv files
  \item \code{\link{tidy_data}} for transforming data into a tidy data set
  \item \code{\link{summarize_data}} for aggregating and summarizing data
}
}

