% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_dates}
\alias{format_dates}
\title{Set the default format for reading date/time variables}
\usage{
format_dates(x)
}
\arguments{
\item{x}{character vector of date/time data}
}
\value{
A \code{\link[readr]{parse_datetime}} object
}
\description{
Set the default format for parsing date/time variables when creating
\code{edwr} classes.
}
\keyword{internal}

