% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.edwr}
\alias{is.blood}
\alias{is.charges}
\alias{is.demographics}
\alias{is.diagnosis}
\alias{is.edwr}
\alias{is.encounters}
\alias{is.events}
\alias{is.icu_assess}
\alias{is.id}
\alias{is.labs}
\alias{is.locations}
\alias{is.measures}
\alias{is.meds_cont}
\alias{is.meds_freq}
\alias{is.meds_home}
\alias{is.meds_sched}
\alias{is.mpp}
\alias{is.mrn}
\alias{is.order_by}
\alias{is.order_detail}
\alias{is.patients}
\alias{is.problems}
\alias{is.procedures}
\alias{is.radiology}
\alias{is.services}
\alias{is.surgeries}
\alias{is.uop}
\alias{is.vent_settings}
\alias{is.vent_times}
\alias{is.visits}
\alias{is.vitals}
\alias{is.warfarin}
\title{Test edwr-related classes}
\usage{
is.edwr(x)

is.blood(x)

is.charges(x)

is.demographics(x)

is.diagnosis(x)

is.encounters(x)

is.events(x)

is.icu_assess(x)

is.id(x)

is.labs(x)

is.locations(x)

is.measures(x)

is.meds_cont(x)

is.meds_freq(x)

is.meds_home(x)

is.meds_sched(x)

is.mpp(x)

is.mrn(x)

is.order_by(x)

is.order_detail(x)

is.patients(x)

is.problems(x)

is.procedures(x)

is.radiology(x)

is.services(x)

is.surgeries(x)

is.uop(x)

is.vent_settings(x)

is.vent_times(x)

is.visits(x)

is.vitals(x)

is.warfarin(x)
}
\arguments{
\item{x}{object which may have an edwr class type}
}
\description{
Takes an R object and checks for an edwr class type.
}

