% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{is.labs}
\alias{is.labs}
\title{Test edwr-related classes}
\usage{
is.labs(x)
}
\arguments{
\item{x}{object which may have an edwr class type}
}
\description{
Takes an R object and checks for an edwr class type.
}

