% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{keep_class}
\alias{keep_class}
\title{Keep edwr class assignments}
\usage{
keep_class(x, y)
}
\arguments{
\item{x}{tibble with edwr class(es)}

\item{y}{tibble as returned by dplyr manipulation function}
}
\value{
tibble
}
\description{
Keep edwr class assignments
}
\keyword{internal}

