% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labs.R
\docType{data}
\name{labs}
\alias{labs}
\title{Sample lab result data}
\format{A data frame with 15 rows and 4 variables
\describe{
  \item{pie.id}{encounter identifier}
  \item{lab.datetime}{date / time the lab was drawn at}
  \item{lab}{name of the lab drawn}
  \item{lab.result}{result of the lab, reported as a character}
}}
\usage{
labs
}
\description{
A dataset containing results for coagulation labs
}
\keyword{datasets}

