% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warfarin.R
\name{make_indications}
\alias{make_indications}
\title{Make warfarin indications}
\usage{
make_indications(x)
}
\arguments{
\item{x}{A data frame with warfarin indications}
}
\value{
A data frame
}
\description{
\code{make_indications} tidy warfarin indication data
}
\details{
This function takes a data frame with the raw warfarin indication as input
into the EMR and returns a tidy data frame with the indication categorized.
It tidies up non-standard indications before returning the values.
}

