% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warfarin_details.R
\name{make_inr_ranges}
\alias{make_inr_ranges}
\title{Make INR therapeutic ranges}
\usage{
make_inr_ranges(x)
}
\arguments{
\item{x}{A data frame with INR ranges}
}
\value{
A data frame
}
\description{
\code{make_inr_ranges} tidy INR ranges data
}
\details{
This function takes a data frame with the goal INR range as input into the
EMR and returns a tidy data frame with the low and high goal values. It
tidies up string errors (typos, incorrect ranges, etc.) before returning the
values.
}
\examples{
print(head(
  make_inr_ranges(warfarin)
))

}
