% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_classes.R
\name{med_class_lookup}
\alias{med_class_lookup}
\title{Lookup medication class by medication generic name}
\usage{
med_class_lookup(meds)
}
\arguments{
\item{meds}{A character vector of medication names}
}
\value{
A data frame with columns: med.class and med.name
}
\description{
\code{med_class_lookup} takes a vector of medication generic names and
returns the classes which contain those medications
}
\details{
This function takes a character vector of medication generic names and
returns a data frame with all of the medication classes that contain those
medications. The data frame will contain two columns: med.class and med.name.
}
