% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med_classes.R
\name{med_lookup}
\alias{med_lookup}
\title{Lookup medications by medication class}
\usage{
med_lookup(med_class)
}
\arguments{
\item{med_class}{A character vector of medication classes}
}
\value{
A data frame with columns: med.class and med.name
}
\description{
\code{med_lookup} takes a vector of medication classes and returns all
medications in those classes
}
\details{
This function takes a character vector of medication classes and returns a
data frame with all of the medications (by generic name) contained in the
medication class. The data frame will contain two columns: med.class and
med.name.

Medication class data comes from the Enterprise Data Warehouse.
}
