% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_perctime.R
\name{perctime}
\alias{perctime}
\title{Calculate percent time above or below a threshold}
\usage{
perctime(x, ..., thrshld)
}
\arguments{
\item{x}{data_frame}

\item{...}{optional grouping variables}

\item{thrshld}{list of criteria}
}
\value{
data_frame
}
\description{
Calculate percent time above or below a threshold
}
\keyword{internal}
