% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data2}
\alias{read_data2}
\title{Read data, automatically assign names and format dates}
\usage{
read_data2(data.dir, file.name, edw = TRUE, recursive = FALSE,
  skip = 0)
}
\arguments{
\item{data.dir}{A character string with the name of the directory containing
the data files}

\item{file.name}{A character string with name of data file or pattern to
match}

\item{edw}{logical. TRUE if the data came from EDW, FALSE for MBO}

\item{recursive}{logical. Should the listing recurse into directories?}

\item{skip}{Number of lines to skip before reading data.}
}
\value{
A data frame of class \code{tbl_edwr}
}
\description{
Read data, automatically assign names and format dates
}
