% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_edw_data}
\alias{read_edw_data}
\title{Read EDW data from csv files}
\usage{
read_edw_data(data.dir, file.name, type = NULL, check.distinct = TRUE,
  include.pts = NULL)
}
\arguments{
\item{data.dir}{A character string with the name of the directory containing
the data files}

\item{file.name}{A character string with name of data file or pattern to
match}

\item{type}{An optional character string indicating type of data being
tidied; uses file.name if no value is provided}

\item{check.distinct}{An optional logical, calls
\code{\link[dplyr]{distinct}} on the imported data if \code{TRUE}}

\item{include.pts}{If not NULL, the returned data frame will be limited to
only the included patients}
}
\value{
A data frame, as an \code{edwr} class type if type is not skip
}
\description{
\code{read_edw_data} takes a directory and file name and reads in all
matching csv files and binds them together into a data frame
}
\details{
This function takes a directory and file name and reads in all matching csv
files and binds them together into a data frame using
\code{\link[readr]{read_csv}} from the readr package.

Valid options for type include:

\itemize{
  \item skip (read in columns without renaming)
  \item blood (blood product administration)
  \item charges (charges by CDM code)
  \item demographics (patient demographics)
  \item diagnosis (ICD-9/10-CM codes)
  \item encounters (all encounters by person)
  \item events (generic clinical events)
  \item home_meds (home / discharge medication lists)
  \item icu_assess (RASS, GCS, etc.)
  \item id (encounter ID, FIN, person ID)
  \item labs (lab results)
  \item measures (height, weight)
  \item meds_cont (continuous medication administration)
  \item meds_sched (intermittent medication administration)
  \item meds_freq (intermittent meds with order frequency)
  \item mpp (MPP used to order medication)
  \item mrn (medical record number)
  \item order_by (ordering provider data)
  \item order_detail (medication order data)
  \item patients (patients identified by CDM, diagnosis codes, or unit admission)
  \item problems (problem list from EMR)
  \item procedures9 (ICD-9 procedure codes)
  \item procedures10 (ICD-10 procedure codes)
  \item radiology (radiology orders)
  \item services (list of patient's primary medical services)
  \item surgeries (list of surgeries from surgery database)
  \item uop (urine output)
  \item vent_settings (FIO2, PaO2, etc.)
  \item vent_start (vent start/stop dates/times)
  \item visits (arrival, admit, discharge data)
  \item vitals (vital signs)
  \item warfarin (warfarin duration, indication, goal)
}
}
\examples{
x <- read_edw_data(
  data.dir = paste0(system.file(package = "edwr", "extdata")),
  file.name = "test_demographics.csv",
  type = "demographics"
)

str(x)

}
\seealso{
\code{\link[readr]{read_csv}}
}

