% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr_verbs.R
\name{reclass}
\alias{reclass}
\title{Keep edwr class assignments}
\usage{
reclass(x, result)
}
\arguments{
\item{x}{tibble with tbl_edwr class(es)}

\item{result}{tibble as returned by dplyr manipulation function}
}
\value{
tibble
}
\description{
Keep edwr class assignments
}
