% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_id_quo}
\alias{set_id_quo}
\title{Set the name of the id field based on data source as a quosure}
\usage{
set_id_quo(x)
}
\arguments{
\item{x}{A tibble with an attribute of "data"}
}
\value{
character vector with the id name
}
\description{
Sets the id name to pie.id for EDW data and millennium.id for MBO data
}
\keyword{internal}
