% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_timezone}
\alias{set_timezone}
\title{Set timezone based on data source}
\usage{
set_timezone(x)
}
\arguments{
\item{x}{A tibble with an attribute of "data"}
}
\value{
character vector indicating the timezone
}
\description{
Sets the timezone to US/Central for EDW data, and UTC for MBO data
}
\keyword{internal}
