% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\alias{summarize_data.default}
\alias{summarize_data.meds_cont}
\alias{summarize_data.meds_inpt}
\alias{summarize_data.meds_home}
\alias{summarize_data.meds_sched}
\alias{summarize_data.labs}
\alias{summarize_data.vitals}
\title{Summary calculations for serial measurement data}
\usage{
summarize_data(x, ...)

\method{summarize_data}{default}(x, ...)

\method{summarize_data}{meds_cont}(x, units = "hours", ...)

\method{summarize_data}{meds_inpt}(x, units = "hours", ...)

\method{summarize_data}{meds_home}(x, ref, pts = NULL, home = TRUE, ...)

\method{summarize_data}{meds_sched}(x, units = "hours", ...)

\method{summarize_data}{labs}(x, units = "hours", ...)

\method{summarize_data}{vitals}(x, units = "hours", ...)
}
\arguments{
\item{x}{A data frame with continuous data}

\item{...}{additional arguments passed on to individual methods}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}

\item{ref}{A data frame with three columns: name, type, and group. See
details below.}

\item{pts}{An optional data frame with a column pie.id including all patients
in study}

\item{home}{A logical, if TRUE (default) look for home medications,
otherwise look for discharge prescriptions}
}
\value{
A data frame
}
\description{
\code{summarize_data} summarizes serial measurement  data
}
\details{
This function takes a data frame with serial measurement  data (e.g.,
medications, lab results) and produces a data frame with summary data for
each patient.

For continuous medications, the calculations include: first rate, last rate,
minimum rate, maximum rate, AUC, time-weighted average rate, total infusion
duration, total infusion running time, and cumulative dose.

For labs and intermittent medications, the calculations include: first, last,
median, maximum, minimum, AUC, and time-weighted average result.

The data frame passed to \code{ref} should contain three character
  columns: name, type, and group. The name column should contain either
  generic medication names or medication classes. The type column should
  specify whether the value in name is a "class" or "med". The group column
  should specify whether the medication is a continous ("cont") or scheduled
  ("sched") medication.
}
\examples{
# make a reference data frame for tidying meds
ref <- tibble::tibble(
  name = c("heparin", "warfarin", "antiplatelet agents"),
  type = c("med", "med", "class"),
  group = c("cont", "sched", "sched")
)

# tidy continuous medications and calculate runtime
x <- tidy_data(meds_cont, ref, meds_sched)
x <- calc_runtime(x)

# pass runtime data to summarize
print(head(
  summarize_data(x)
))

# make a reference data frame for tidying meds
ref <- tibble::tibble(
  name = c("heparin", "warfarin", "antiplatelet agents"),
  type = c("med", "med", "class"),
  group = c("cont", "sched", "sched")
)

# tidy home medications
print(head(
  summarize_data(meds_home, ref)
))

# return all patients, even if they do not have any of the desired home meds
pts <- dplyr::distinct(meds_home, pie.id)
print(head(
  summarize_data(meds_home, ref, pts = pts)
))

# return discharge prescriptions instead of home meds
print(head(
  summarize_data(meds_home, ref, pts = pts, home = FALSE)
))

}
