% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summary_fun}
\alias{summary_fun}
\title{Performs the summarize operation}
\usage{
summary_fun(df, grp, dt_col, val_col)
}
\arguments{
\item{grp}{grouping quosures}

\item{dt_col}{datetime column}

\item{val_col}{numeric value column}

\item{x}{tibble}
}
\value{
tibble
}
\description{
Performs the summarize operation
}
\keyword{internal}
