% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data.edw_labs}
\alias{tidy_data.edw_labs}
\title{Tidy lab results}
\usage{
\method{tidy_data}{edw_labs}(x, censor = TRUE)
}
\arguments{
\item{x}{A data frame with raw lab results}

\item{censor}{A logical, if TRUE will add a column indicating the data was
censored (default)}
}
\value{
A data frame
}
\description{
\code{tidy_labs} transforms lab result data into a tidy format
}
\details{
This function takes a data frame with lab results and returns a tidy data
frame. Results will be converted to numeric values and censored data will be
indicated.
}

