% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_fun}
\alias{tidy_fun}
\title{Title}
\usage{
tidy_fun(x, group, ref = NULL)
}
\arguments{
\item{x}{tibble}

\item{group}{string indicating cont or sched}

\item{ref}{}
}
\value{
tibble
}
\description{
Title
}
\keyword{internal}
