% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_icd}
\alias{tidy_icd}
\title{Tidy ICD9/10 diagnosis codes}
\usage{
tidy_icd(raw.data, ref.data, icd10 = FALSE, patients = NULL)
}
\arguments{
\item{raw.data}{A data frame with all patient diagnosis codes}

\item{ref.data}{A data frame with the desired diagnosis codes}

\item{icd10}{A logical indicating whether to use ICD-10 codes or ICD-9 codes
(default)}

\item{patients}{An optional data frame with a column pie.id including all
patients in study}
}
\value{
A data frame
}
\description{
\code{tidy_icd} determines which patients have the desired ICD9/10 diagnosis
}
\details{
This function is called by \code{\link{tidy_data}} with a \code{type}
argument of \code{icd9} or \code{icd10}.

The function takes a data frame with reference diagnosis codes and a data
frame with all patient diagnosis codes, and returns a data frame with a
column for each disease state for each patient.
}
\seealso{
\code{\link{tidy_data}}
}

