% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_runtime.R
\name{calc_runtime}
\alias{calc_runtime}
\alias{calc_runtime.default}
\alias{calc_runtime.meds_cont}
\alias{calc_runtime.meds_inpt}
\alias{calc_runtime.meds_sched}
\alias{calc_runtime.labs}
\alias{calc_runtime.events}
\alias{calc_runtime.vitals}
\title{Calculate the running time for serial measurement data}
\usage{
calc_runtime(x, ...)

\method{calc_runtime}{default}(x, ...)

\method{calc_runtime}{meds_cont}(x, drip.off = 12, no.doc = 24,
  units = "hours", ...)

\method{calc_runtime}{meds_inpt}(x, drip.off = 12, no.doc = 24,
  units = "hours", cont = TRUE, ...)

\method{calc_runtime}{meds_sched}(x, units = "hours", ...)

\method{calc_runtime}{labs}(x, units = "hours", ...)

\method{calc_runtime}{events}(x, units = "hours", ...)

\method{calc_runtime}{vitals}(x, units = "hours", ...)
}
\arguments{
\item{x}{A data frame with serial measurement data}

\item{...}{additional arguments passed on to individual methods}

\item{drip.off}{An optional numeric indicating the number of hours a
medication infusion should be off to count as a new infusion, defaults to
12 hours}

\item{no.doc}{An optional numeric indicating the number of hours without
documentation which will be used to indicate a drip has ended, defaults to
24 hours}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}

\item{cont}{A logical, if TRUE (default), treat the medications as continuous
when summarizing}
}
\value{
A data frame
}
\description{
\code{calc_runtime} calculates the duration of time at current value and
total run time
}
\details{
This function takes a data frame with serial measurement data (e.g.,
continuous medications or repeated lab data) and produces a data frame with
the with the duration of time at each value and cumulative run time.

This could be used to then calculate the AUC or to summarize the continuous
data.

For continuous medications, the data will be grouped into distinct
  sets of infusions, for patients who may have been restarted on the drip one
  or more times. Use the \code{drip.off} argument to modify the criteria for
  determining distinct infusions.
}
\examples{
# make a reference data frame for tidying meds
ref <- tibble::tibble(
  name = c("heparin", "warfarin", "antiplatelet agents"),
  type = c("med", "med", "class"),
  group = c("cont", "sched", "sched")
)

# tidy continuous medications; will keep only heparin drips
x <- tidy_data(meds_cont, meds_sched, ref)

# calculate the runtime for continuous heparin infusion
print(head(
  calc_runtime(x)
))

}
