% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_pregnant.R
\name{check_pregnant}
\alias{check_pregnant}
\alias{check_pregnant.default}
\alias{check_pregnant.diagnosis}
\alias{check_pregnant.labs}
\title{Identify pregnant patients}
\usage{
check_pregnant(x, ...)

\method{check_pregnant}{default}(x, ...)

\method{check_pregnant}{diagnosis}(x, ...)

\method{check_pregnant}{labs}(x, ...)
}
\arguments{
\item{x}{A data frame of either class diagnosis or labs}

\item{...}{additional arguments passed on to individual methods}
}
\value{
A data frame
}
\description{
\code{check_pregnant} takes data frames with either diagnosis codes or labs
and returns a data frame with identifiers of pregnant patients
}
\details{
This function takes a data frame of either class diagnosis or labs and
returns a data frame with a list of all patients who are pregnant. This can
be used to exclude patients from research studies.
}

