% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_dates}
\alias{format_dates}
\title{Convert date/time variables to POSIXct}
\usage{
format_dates(x, date_col)
}
\arguments{
\item{x}{character vector of date/time data}

\item{date_col}{character of column names to be converted}
}
\value{
A tibble
}
\description{
Converts date/time variables to POSIXct and adjusts to US/Central timezone.
}
\keyword{internal}

