% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{format_dates}
\alias{format_dates}
\title{Set the default format for reading date/time variables}
\usage{
format_dates(x)
}
\arguments{
\item{x}{character vector of date/time data}
}
\value{
A readr::collector object#'
}
\description{
Set the default format for reading date/time variables
}
\keyword{internal}

