% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_perctime.R
\name{perctime}
\alias{perctime}
\title{Calculate percent time above or below a threshold}
\usage{
perctime(x, thrshld, vars)
}
\arguments{
\item{x}{data_frame}

\item{thrshld}{list of criteria}

\item{vars}{character vector with columns for grouping}
}
\value{
data_frame
}
\description{
Calculate percent time above or below a threshold
}
\keyword{internal}

