% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data from csv files}
\usage{
read_data(data.dir, file.name, type = NULL, check.distinct = TRUE,
  include.pts = NULL)
}
\arguments{
\item{data.dir}{A character string with the name of the directory containing
the data files}

\item{file.name}{A character string with name of data file or pattern to
match}

\item{type}{An optional character string indicating type of data being
tidied; uses file.name if no value is provided}

\item{check.distinct}{An optional logical, calls
\code{\link[dplyr]{distinct}} on the imported data if \code{TRUE}}

\item{include.pts}{If not NULL, the returned data frame will be limited to
only the included patients}
}
\value{
A data frame, as an \code{edwr} class type if type is not skip
}
\description{
\code{read_data} takes a directory and file name and reads in all
matching csv files and binds them together into a data frame
}
\details{
This function takes a directory and file name and reads in all matching csv
files and binds them together into a data frame using
\code{\link[readr]{read_csv}} from the readr package.

Valid options for \code{type} are listed below, along with the names of the
EDW Standard Project Queries which correspond to each \code{type}.

\tabular{lcl}{
  \strong{type Argument} \tab --- \tab \strong{EDW Query Name}\cr
  skip \tab \tab query independent; reads in columns without renaming\cr
  blood \tab \tab Blood Products\cr
  charges \tab \tab Charges - [All, Department Prompt]\cr
  demographics \tab \tab Demographics\cr
  diagnosis \tab \tab Diagnosis Codes (ICD-9/10-CM) - All\cr
  encounters \tab \tab Encounters - by Person ID\cr
  events \tab \tab Clinical Events - Prompt\cr
  icu_assess \tab \tab ICU Assessments (CAM-ICU, GCS, RASS)\cr
  id \tab \tab Identifiers - by FIN\cr
  \tab \tab Identifiers - by MRN\cr
  \tab \tab Identifiers - by PowerInsight Encounter Id\cr
  labs \tab \tab Labs - ABG\cr
  \tab \tab Labs - CBC\cr
  \tab \tab Labs - Chemistry\cr
  \tab \tab Labs - Coags\cr
  \tab \tab Labs - HIT\cr
  \tab \tab Labs - LFTs\cr
  \tab \tab Labs - Pregnancy\cr
  \tab \tab Labs - Renal\cr
  \tab \tab Urine Drug Screen\cr
  \tab \tab Vomiting Output\cr
  locations \tab \tab Location History\cr
  measures \tab \tab Measures (Height and Weight)\cr
  meds_cont \tab \tab Medications - Inpatient Continuous - [All, Prompt]\cr
  meds_freq \tab \tab Medications - Inpatient Intermittent with Frequency - [All,
    Prompt]\cr
  meds_home \tab \tab Medications - Home and Discharge - [All, Prompt]\cr
  meds_sched \tab \tab Medications - Inpatient Intermittent - [All, Prompt]\cr
  mpp \tab \tab MPP's Ordered - [All, Order Prompt]\cr
  mrn \tab \tab Identifiers - Person by MRN\cr
  order_by \tab \tab Orders - Prompt\cr
  order_detail \tab \tab Orders - from Clinical Event Id - Prompt\cr
  patients \tab \tab Patients - by CDM charge - HealthQuest - [Location Prompt,
    System]\cr
  \tab \tab Patients - by ICD-9/10-CM Diagnosis Code\cr
  \tab \tab Patients - by Medication\cr
  \tab \tab Patients - by Unit Admission\cr
  problems \tab \tab Problem List\cr
  procedures9 \tab \tab Procedure Codes (ICD-9-CM) - [All, Prompt]\cr
  procedures10 \tab \tab Procedure Codes (ICD-10-PCS) - [All, Prompt]\cr
  radiology \tab \tab Radiology Reports\cr
  services \tab \tab Service History\cr
  surgeries \tab \tab Surgeries\cr
  uop \tab \tab Urine output\cr
  vent_settings \tab \tab Ventilator Data - Settings\cr
  vent_start \tab \tab Ventilator Data - Start and Stop\cr
  visits \tab \tab Visit Data\cr
  vitals \tab \tab Vitals\cr
  warfarin \tab \tab Warfarin Information\cr
}
}
\examples{
x <- read_data(
  data.dir = paste0(system.file(package = "edwr", "extdata")),
  file.name = "demographics.csv",
  type = "demographics"
)

str(x)

}
\seealso{
\code{\link[readr]{read_csv}}
}

