% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_edw_data}
\alias{read_edw_data}
\title{Read EDW data from csv files}
\usage{
read_edw_data(data.dir, file.name, type = NULL, check.distinct = TRUE,
  include.pts = NULL)
}
\arguments{
\item{data.dir}{A character string with the name of the directory containing
the data files}

\item{file.name}{A character string with name of data file or pattern to
match}

\item{type}{An optional character string indicating type of data being
tidied; uses file.name if no value is provided}

\item{check.distinct}{An optional logical, calls
\code{\link[dplyr]{distinct}} on the imported data if \code{TRUE}}

\item{include.pts}{If not NULL, the returned data frame will be limited to
only the included patients}
}
\value{
A data frame, as an \code{edwr} class type if type is not skip
}
\description{
\code{read_edw_data} takes a directory and file name and reads in all
matching csv files and binds them together into a data frame
}
\details{
This function takes a directory and file name and reads in all matching csv
files and binds them together into a data frame using
\code{\link[readr]{read_csv}} from the readr package.

Valid options for type include: skip (read in columns without renaming),
blood, charges, demographics, diagnosis, encounters, events, home_meds,
icu_assess, id, labs, locations, measures, meds_cont, meds_cont_detail,
meds_sched, meds_sched_freq, mpp, mrn, orders, order_detail, patients,
problems, procedures_icd9, procedures_icd10, radiology, services, surgeries,
uop, vent_settings, vent_start, visits, vitals, warfarin
}
\examples{
x <- read_edw_data(data.dir = paste0(system.file(package = "edwr", "extdata")),
                     file.name = "test_demographics.csv",
                     type = "demographics")
str(x)

}
\seealso{
\code{\link[readr]{read_csv}}
}

