% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_edw_data}
\alias{read_edw_data}
\title{Read EDW data from csv files}
\usage{
read_edw_data(data.dir, file.name, type = NA, check.distinct = TRUE)
}
\arguments{
\item{data.dir}{A character string with the name of the directory containing
the data files}

\item{file.name}{A character string with name of data file or pattern to
match}

\item{type}{An optional character string indicating type of data being tidied}

\item{check.distinct}{An optional logical, calls
\code{\link[dplyr]{distinct}} on the imported data if \code{TRUE}}
}
\value{
A data frame (tbl_df)
}
\description{
\code{read_edw_data} takes a directory and file name and reads in all
matching csv files and binds them together into a data frame
}
\details{
This function takes a directory and file name and reads in all matching csv
files and binds them together into a data frame using
\code{\link[readr]{read_csv}} from the readr package.

Valid options for type include: blood, charges, demographics, diagnosis*,
encounters, events, home_meds, icd9, icd10, icu_assess, id, labs, locations,
measures, meds_continuous, meds_sched, meds_sched_freq, mpp, mrn, orders,
patients, problems, procedures**, procedures_icd9, procedures_icd10,
radiology, services, surgeries, uop, vent_settings, vent_start, visits,
vitals, warfarin

* diagnosis option is deprecated; use icd9 or icd10 instead

** procedures option is deprecated; use procedures_icd9 or procedures_icd10
instead
}
\seealso{
\code{\link[readr]{read_csv}}
}

