% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{set_edwr_class}
\alias{as.blood}
\alias{as.charges}
\alias{as.demographics}
\alias{as.diagnosis}
\alias{as.drg}
\alias{as.encounters}
\alias{as.events}
\alias{as.icu_assess}
\alias{as.id}
\alias{as.labs}
\alias{as.locations}
\alias{as.measures}
\alias{as.meds_cont}
\alias{as.meds_freq}
\alias{as.meds_home}
\alias{as.meds_sched}
\alias{as.mpp}
\alias{as.mrn}
\alias{as.order_action}
\alias{as.order_by}
\alias{as.order_detail}
\alias{as.order_info}
\alias{as.order_timing}
\alias{as.patients}
\alias{as.problems}
\alias{as.procedures}
\alias{as.radiology}
\alias{as.services}
\alias{as.surgeries}
\alias{as.tbl_edwr}
\alias{as.uop}
\alias{as.vent_settings}
\alias{as.vent_times}
\alias{as.visits}
\alias{as.vitals}
\alias{as.warfarin}
\alias{set_edwr_class}
\alias{tbl_edwr}
\title{Construct edwr data types}
\usage{
tbl_edwr(x)

as.tbl_edwr(x)

as.blood(x)

as.charges(x)

as.demographics(x, varnames = NULL, extras = NULL)

as.diagnosis(x, varnames = NULL, extras = NULL)

as.drg(x)

as.encounters(x)

as.events(x, varnames = NULL, extras = NULL)

as.icu_assess(x)

as.id(x)

as.labs(x)

as.locations(x)

as.measures(x)

as.meds_cont(x)

as.meds_freq(x)

as.meds_home(x)

as.meds_sched(x)

as.mrn(x)

as.mpp(x)

as.order_action(x, varnames = NULL, extras = NULL)

as.order_by(x)

as.order_detail(x)

as.order_info(x)

as.order_timing(x)

as.patients(x, varnames = NULL, extras = NULL)

as.problems(x)

as.procedures(x)

as.radiology(x)

as.services(x)

as.surgeries(x)

as.uop(x)

as.vent_settings(x)

as.vent_times(x)

as.visits(x)

as.vitals(x)

as.warfarin(x)
}
\arguments{
\item{x}{object to set class \code{tbl_edwr}}

\item{varnames}{named character list, where the name is the desired variable
name; overrides default names}

\item{extras}{named character list; if given, will append to default varnames}
}
\description{
Takes an R object and sets class to an edwr type.
}
\keyword{internal}

