% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{set_edwr_class}
\alias{as.demographics}
\alias{as.edwr}
\alias{as.labs}
\alias{edwr}
\alias{set_edwr_class}
\title{Construct edwr data types}
\usage{
edwr(x)

as.edwr(x)

as.demographics(x)

as.labs(x)
}
\arguments{
\item{x}{object to set class \code{edwr}}
}
\description{
Takes an R object and sets class to an edwr type.
}
\keyword{internal}

