% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{set_edwr_class}
\alias{as.labs}
\alias{labs}
\alias{set_edwr_class}
\title{Construct edwr data types}
\usage{
labs(x)

as.labs(x)
}
\arguments{
\item{x}{object to set class \code{labs}}
}
\description{
Takes an R object and sets class to an edwr type.
}
\keyword{internal}

