% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\name{set_edwr_class}
\alias{as.blood}
\alias{as.charges}
\alias{as.demographics}
\alias{as.diagnosis}
\alias{as.edwr}
\alias{as.encounters}
\alias{as.events}
\alias{as.icu_assess}
\alias{as.id}
\alias{as.labs}
\alias{as.locations}
\alias{as.measures}
\alias{as.meds_cont}
\alias{as.meds_freq}
\alias{as.meds_home}
\alias{as.meds_sched}
\alias{as.services}
\alias{as.vent_times}
\alias{as.visits}
\alias{as.warfarin}
\alias{edwr}
\alias{set_edwr_class}
\title{Construct edwr data types}
\usage{
edwr(x)

as.edwr(x)

as.blood(x)

as.charges(x)

as.demographics(x)

as.diagnosis(x)

as.encounters(x)

as.events(x)

as.icu_assess(x)

as.id(x)

as.labs(x)

as.locations(x)

as.measures(x)

as.meds_cont(x)

as.meds_freq(x)

as.meds_home(x)

as.meds_sched(x)

as.services(x)

as.vent_times(x)

as.visits(x)

as.warfarin(x)
}
\arguments{
\item{x}{object to set class \code{edwr}}
}
\description{
Takes an R object and sets class to an edwr type.
}
\keyword{internal}

