% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_data.R
\name{summarize_data}
\alias{summarize_data}
\alias{summarize_data.default}
\alias{summarize_data.labs}
\alias{summarize_data.meds_cont}
\alias{summarize_data.meds_sched}
\title{Summary calculations for continuous data}
\usage{
summarize_data(x, ...)

\method{summarize_data}{default}(x, ...)

\method{summarize_data}{meds_cont}(x, units = "hours", ...)

\method{summarize_data}{meds_sched}(x, units = "hours", ...)

\method{summarize_data}{labs}(x, units = "hours", ...)
}
\arguments{
\item{x}{A data frame with continuous data}

\item{units}{An optional character string specifying the time units to use in
calculations, default is hours}
}
\value{
A data frame
}
\description{
\code{summarize_data} summarizes continuous data
}
\details{
This function takes a data frame with continuous data (e.g., medications, lab
results) and produces a data frame with summary data for each patient.

For continuous medications, the calculations include: first rate, last rate,
minimum rate, maximum rate, AUC, time-weighted average rate, total infusion
duration, total infusion running time, and cumulative dose.

For labs, the calculations include: first, last, median, maximum, minimum,
AUC, and time-weighted average result.
}
\examples{
# make a reference data frame for tidying meds
ref <- tibble::tibble(
  name = c("heparin", "warfarin", "antiplatelet agents"),
  type = c("med", "med", "class"),
  group = c("cont", "sched", "sched")
)

# tidy continuous medications and calculate runtime
x <- tidy_data(meds_cont, ref, meds_sched)
x <- calc_runtime(x)

# pass runtime data to summarize
print(head(
  summarize_data(x)
))

}

