% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_data}
\alias{tidy_data}
\alias{tidy_data.default}
\alias{tidy_data.labs}
\alias{tidy_data.meds_cont}
\alias{tidy_data.meds_home}
\alias{tidy_data.meds_sched}
\title{Tidy data}
\usage{
tidy_data(x, ...)

\method{tidy_data}{default}(x, ...)

\method{tidy_data}{labs}(x, censor = TRUE, ...)

\method{tidy_data}{meds_cont}(x, ref, sched, ...)

\method{tidy_data}{meds_sched}(x, ref, ...)

\method{tidy_data}{meds_home}(x, ref, pts = NULL, home = TRUE, ...)
}
\arguments{
\item{x}{A data frame with an edw class type}

\item{...}{additional arguments passed on to individual methods}

\item{censor}{A logical, if TRUE will add a column indicating the data was
censored (default)}

\item{ref}{A data frame with three columns: name, type, and group. See
details below.}

\item{sched}{A data frame with intermittent medications}

\item{pts}{An optional data frame with a column pie.id including all patients
in study}

\item{home}{A logical, if TRUE (default) look for home medications,
otherwise look for discharge prescriptions}
}
\description{
\code{tidy_data} transforms raw EDW data into a tidy format
}
\details{
This is an S3 generic function for tidying EDW data read in using
\code{\link{read_data}}. The function invokes the appropriate method based on
the type of data being transformed (i.e., lab results, medication data,
etc.).

The data frame passed to \code{ref} should contain three character columns:
name, type, and group. The name column should contain either generic
medication names or medication classes. The type column should specify
whether the value in name is a "class" or "med". The group column should
specify whether the medication is a continous ("cont") or scheduled ("sched")
medication.
}
\examples{
# tidy lab data; non-numeric results will be converted to NA
suppressWarnings(head(
  tidy_data(labs)
))

# tidy labs without marking censored data (will be converted to NA's)
suppressWarnings(head(
  tidy_data(labs, censor = FALSE)
))

# make a reference data frame for tidying meds
ref <- tibble::data_frame(
  name = c("heparin", "warfarin", "antiplatelet agents"),
  type = c("med", "med", "class"),
  group = c("cont", "sched", "sched")
)

# tidy continuous medications; will keep only heparin drips
heparin <- tidy_data(meds_cont, ref, meds_sched)
head(heparin)

# tidy intermittent medications; will keep warfarin and antiplatelet agents
sched <- tidy_data(meds_sched, ref)
head(sched)

# tidy home medications
hm <- tidy_data(meds_home, ref)
head(hm)

# return all patients, even if they do not have any of the desired home meds
pts <- dplyr::distinct(labs, pie.id)
hm.all <- tidy_data(meds_home, ref, pts = pts)

# return discharge prescriptions instead of home meds
dc.all <- tidy_data(meds_home, ref, pts = pts, home = FALSE)

}

